// src/features/main/parking/blocks/guest_vehicle_info_input_box/types.ts
export type GuestVehicleInfo = {
    plateNumber: string;
    phoneNumber: string;
    visitingPerpose: string;
};

export type TouchedCoordinate = {
    x: number;
    y: number;
};

export type GuestVehicleValidationResult = {
    plateNumber: boolean;
    phoneNumber: boolean;
    visitingPerpose: boolean;
};

export type GuestVehicleInfoInputBoxProps = {
    initialVehicleInfo?: GuestVehicleInfo;
    onValidation?(validationResult: GuestVehicleValidationResult): void;
    onChangeGuestVehicleInfo(vehicleInfo: GuestVehicleInfo): void;
};